requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pycryptodome>=3.19.0
